+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Slapfish not being recognized.
 - Fishslap and Slapfish chat messages not being themed correctly.
 - High Contrast Theme not modifying rank color in tab header.
+- Message Preview showing empty chat message on unknown servers.
 
 ## 0.2.0: 2025-04-05